send reminder for overdue interviews

+++ src/main/kotlin/com/mindbridgehealth/footing/service/ScheduledInterviewInitiationTask.kt
@@ -15,6 +15,10 @@ import java.net.URLEncoder
 import java.nio.charset.StandardCharsets
 import java.sql.Timestamp
 import java.time.Instant
+import java.time.LocalTime
+import java.time.ZoneId
+import java.time.ZonedDateTime
+import java.time.temporal.TemporalField
 
 @Component
 class ScheduledInterviewInitiationTask(
@@ -36,13 +40,18 @@ class ScheduledInterviewInitiationTask(
     fun performTask() {
         val scheduledInterviews = scheduledInterviewRepository.findAllByScheduledTime(Timestamp.from(Instant.now().plusSeconds(60)))
         logger.debug("Looking for scheduled interviews")
-        scheduledInterviews.forEach { scheduledInterviewEntity ->
-            processScheduledInterview(scheduledInterviewEntity)
-        }
+        scheduledInterviews
+            .filter { scheduledInterviewEntity -> !scheduledInterviewEntity.linkSent!! }
+            .forEach { scheduledInterviewEntity -> processScheduledInterview(scheduledInterviewEntity) }
+
+        logger.debug("Looking for overdue scheduled interviews")
+        val potentialOverdueInterviews = scheduledInterviewRepository.findAllByLinkSent(true)
+        potentialOverdueInterviews
+            .filter { scheduledInterviewEntity -> isTimeToSendReminder(scheduledInterviewEntity) }
+            .forEach { scheduledInterviewEntity -> processScheduledInterview(scheduledInterviewEntity) }
     }
 
     private fun processScheduledInterview(scheduledInterviewEntity: ScheduledInterviewEntity) {
-        logger.debug("Reminding Interview: ${scheduledInterviewEntity.name}")
         val interview = scheduledInterviewEntity.interview?.id?.let { interviewService.findInterviewEntityById(it) }
         val interviewQuestions = interview?.id?.let { interviewQuestionService.findEntitiesByInterviewId(it) }
         if(interviewQuestions.isNullOrEmpty()) {
@@ -62,16 +71,34 @@ class ScheduledInterviewInitiationTask(
         if (number == null) {
             logAndThrow("Unable to find mobile number for Storyteller ${interview.storyteller?.id} scheduled interview: ${scheduledInterviewEntity.name} Interview Id: ${scheduledInterviewEntity.interview?.id}")
         }
-        val questionString =
-            "Here is your question for your MindBridge Health interview: $interviewQuestionText"
+
         val encodedQuestion =
             URLEncoder.encode(interviewQuestionText, StandardCharsets.UTF_8.toString())
-        val url =
-            "http://app.mindbridgehealth.com/interview.html?cid=6w7x8y9z&mid=1a2b3c4b&rid=12345&rtel=$number&question=$encodedQuestion&interview_id=$interviewAltId&interview_question_id=$interviewQuestionId"
+        val url = "http://app.mindbridgehealth.com/interview.html?cid=6w7x8y9z&mid=1a2b3c4b&rid=12345&rtel=$number&question=$encodedQuestion&interview_id=$interviewAltId&interview_question_id=$interviewQuestionId"
 
-        smsNotificationService.sendInterviewLink(number!!, name, questionString, url, interviewAltId, interviewQuestionId)
+        if(!scheduledInterviewEntity.linkSent!!) {
+            val questionString =
+                "Here is your question for your MindBridge Health interview: $interviewQuestionText"
+            logger.debug("Sending initial interview: ${scheduledInterviewEntity.name}")
+            smsNotificationService.sendInterviewLink(number!!, name, questionString, url, interviewAltId, interviewQuestionId)
 
-        scheduledInterviewRepository.delete(scheduledInterviewEntity)
+            scheduledInterviewEntity.linkSent = true
+            scheduledInterviewRepository.save(scheduledInterviewEntity)
+        } else {
+            logger.debug("Reminding interview: ${scheduledInterviewEntity.name}")
+            val message = "Hello $name! \uD83C\uDF1F Your unique memories are like stars waiting to shine. \uD83C\uDF20 We'd love to hear your response to this week's question! Your story is invaluable. Take a moment to share your wisdom. \uD83D\uDCDD\uD83E\uDDE1 http://app.mindbridgehealth.com/interview.html?cid=6w7x8y9z&mid=1a2b3c4b&rid=12345&rtel=$number&question=$encodedQuestion&interview_id=$interviewAltId&interview_question_id=$interviewQuestionId"
+            smsNotificationService.sendMessage(number!!, message)
+            scheduledInterviewRepository.delete(scheduledInterviewEntity)
+        }
+
+    }
+
+    //Todo: This will have to take user timezone into account at the very least probably reminder preferences too
+    fun isTimeToSendReminder(scheduledInterviewEntity: ScheduledInterviewEntity): Boolean {
+//        val currentTime = ZonedDateTime.now(ZoneId.of("UTC"))
+//        val targetTime = ZonedDateTime.of(currentTime.toLocalDate(), LocalTime.of(15, 0), ZoneId.of("America/New_York"))
+//
+        return scheduledInterviewEntity.linkSent!! //&& currentTime.isAfter(targetTime)
     }
 
     private fun logAndThrow(errorMsg: String) {
